Move fixtures to test file
